+++ skills_embed.go
@@ -1,0 +1,13 @@
+package findo
+
+import (
+	"embed"
+	"io/fs"
+)
+
+//go:embed skills/*
+var embeddedSkills embed.FS
+
+func EmbeddedSkills() (fs.FS, error) {
+	return fs.Sub(embeddedSkills, "skills")
+}